Version 0.1.10

+++ src/resources/combat_guild_reputation.rs
@@ -1,4 +1,4 @@
 use bevy::ecs::system::Resource;
 
 #[derive(Resource)]
-pub struct CombatGuildReputation(pub f32);
+pub struct CombatGuildReputation(pub u32);
