+++ src/resources/exploration_guild_reputation.rs
@@ -1,4 +1,4 @@
 use bevy::ecs::system::Resource;
 
 #[derive(Resource)]
-pub struct ExplorationGuildReputation(pub f32);
+pub struct ExplorationGuildReputation(pub u32);
