+++ src/resources/mine_ammunition.rs
@@ -1,4 +1,4 @@
 use bevy::ecs::system::Resource;
 
 #[derive(Resource)]
-pub struct MineAmmunition(pub f32);
+pub struct MineAmmunition(pub u32);
