+++ src/resources/torpedo_ammunition.rs
@@ -1,4 +1,4 @@
 use bevy::ecs::system::Resource;
 
 #[derive(Resource)]
-pub struct TorpedoAmmunition(pub f32);
+pub struct TorpedoAmmunition(pub u32);
